+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog for the [`ghc-typelits-extra`](http://hackage.haskell.org/package/ghc-typelits-extra) package
 
-# 0.2.2
+# 0.2.2 *January 15th 2017*
 * Reduce `Min n (n+1)` to `n`
 * Reduce `Max n (n+1)` to `n+1`
 * Reduce cases like `1 <=? Div 18 6` to `True`